# specs/001-noorestan-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Noorestan Catalog API
  version: 1.0.0-draft
  description: >-
    Contract for the public catalog and protected administration application. There are no cart,
    checkout, payment, customer-account, online-order, inventory, or visitor-inquiry endpoints.
servers:
  - url: /api/v1
...
paths:
...
  /admin/products:
    get:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [] }]
      operationId: listAdminProducts
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ProductStatus' } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Bounded administrator product list }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductWrite' }
      responses:
        '201':
          description: Draft product created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProduct' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /admin/products/{id}:
    get:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [] }]
      operationId: getAdminProduct
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Product in any non-deleted state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProduct' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductWrite' }
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProduct' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: permanentlyDeleteProduct
      description: Permanently deletes an archived product after explicit confirmation.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - name: confirmation
          in: header
          required: true
          schema: { type: string, const: permanently-delete }
      responses:
        '204': { description: Permanently deleted; object cleanup completed or durably queued }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/products/{id}/status:
    post:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: transitionProductStatus
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetStatus]
              properties:
                targetStatus: { $ref: '#/components/schemas/ProductStatus' }
      responses:
        '200':
          description: Transition completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProduct' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/categories:
    get:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [] }]
      operationId: listAdminCategories
      responses:
        '200': { description: Categories with specification definitions and product counts }
    post:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryWrite' }
      responses:
        '201': { description: Category created }
        '400': { $ref: '#/components/responses/ValidationError' }
  /admin/categories/{id}:
    put:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryWrite' }
      responses:
        '200': { description: Category and definition changes saved }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Admin Catalog]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: deleteCategory
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Empty category deleted }
        '409':
          description: Category still contains non-archived products or version is stale
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /owner/accounts:
    get:
      tags: [Owner Accounts]
      security: [{ cookieAuth: [] }]
      operationId: listAdministratorAccounts
      responses:
        '200': { description: Administrator accounts; owner policy required }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Owner Accounts]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: createAdministratorAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdministratorAccountCreate' }
      responses:
        '201': { description: Individual administrator account created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /owner/accounts/{id}/deactivate:
    post:
      tags: [Owner Accounts]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: deactivateAdministratorAccount
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Account deactivated and sessions revoked }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /owner/transfer:
    post:
      tags: [Owner Accounts]
      security: [{ cookieAuth: [], csrfToken: [] }]
      operationId: transferOwnership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newOwnerId]
              properties:
                newOwnerId: { type: string, format: uuid }
      responses:
        '204': { description: Ownership transferred atomically }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-noorestan-session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Opaque concurrency token from the latest representation.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: Authenticated account lacks the required owner policy
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource does not exist or is not public
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Concurrency or lifecycle/integrity conflict
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    ValidationError:
      description: Request validation failed
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ValidationProblem' } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId: { type: string }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties: { type: array, items: { type: string } }
    ProductStatus: { type: string, enum: [draft, published, hidden, archived] }
    Image:
      type: object
      required: [id, altTextFa, variants]
      properties:
        id: { type: string, format: uuid }
        altTextFa: { type: string }
        isPrimary: { type: boolean }
        variants:
          type: array
          items:
            type: object
            required: [url, width, height, mediaType]
            properties:
              url: { type: string, format: uri-reference }
              width: { type: integer }
              height: { type: integer }
              mediaType: { type: string }
...
    AdministratorAccountCreate:
      type: object
      additionalProperties: false
      required: [displayName, email, temporaryPassword]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email, maxLength: 320 }
        temporaryPassword: { type: string, minLength: 12, maxLength: 1024 }
...
    SpecificationValueWrite:
      type: object
      required: [definitionId]
      properties:
        definitionId: { type: string, format: uuid }
        textValue: { type: string }
        numericValue: { type: number }
        booleanValue: { type: boolean }
        choiceId: { type: string, format: uuid }
    ProductWrite:
      type: object
      additionalProperties: false
      required: [categoryId, slug, nameFa, shortDescriptionFa, descriptionFa, features, specifications]
      properties:
        categoryId: { type: string, format: uuid }
        slug: { type: string, minLength: 1, maxLength: 160 }
        nameFa: { type: string, minLength: 1, maxLength: 300 }
        shortDescriptionFa: { type: string, minLength: 1, maxLength: 500 }
        descriptionFa: { type: string, minLength: 1, maxLength: 20000 }
        technicalNotesFa: { type: string, maxLength: 20000 }
        displayOrder: { type: integer, minimum: 0 }
        features: { type: array, items: { type: string, maxLength: 500 }, maxItems: 100 }
        specifications: { type: array, items: { $ref: '#/components/schemas/SpecificationValueWrite' }, maxItems: 100 }
    AdminProduct:
      allOf:
        - $ref: '#/components/schemas/ProductWrite'
        - type: object
          required: [id, status, images, concurrencyToken]
          properties:
            id: { type: string, format: uuid }
            status: { $ref: '#/components/schemas/ProductStatus' }
            images: { type: array, items: { $ref: '#/components/schemas/Image' } }
            concurrencyToken: { type: string }
    SpecificationDefinitionWrite:
      type: object
      required: [key, labelFa, valueType, isRequired, isFilterable, displayOrder]
      properties:
        key: { type: string, maxLength: 100 }
        labelFa: { type: string, maxLength: 200 }
        valueType: { type: string, enum: [text, number, boolean, choice] }
        unitFa: { type: string, maxLength: 50 }
        isRequired: { type: boolean }
        isFilterable: { type: boolean }
        displayOrder: { type: integer, minimum: 0 }
        choices: { type: array, items: { type: string, maxLength: 200 } }
    CategoryWrite:
      type: object
      required: [slug, nameFa, displayOrder, isVisible, specifications]
      properties:
        slug: { type: string, maxLength: 160 }
        nameFa: { type: string, maxLength: 300 }
        descriptionFa: { type: string, maxLength: 5000 }
        displayOrder: { type: integer, minimum: 0 }
        isVisible: { type: boolean }
        specifications: { type: array, items: { $ref: '#/components/schemas/SpecificationDefinitionWrite' } }
...